# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Manabi API",
    "description": "JLPT 专项练习 API。所有时间为 UTC，媒体地址相对于 API 根地址。",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/exams": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "summary": "Exams",
        "operationId": "exams_api_v1_exams_get",
        "parameters": [
          {
            "name": "level",
            "in": "query",
            "required": true,
            "schema": {
              "enum": [
                "N2",
                "N3"
              ],
              "type": "string",
              "title": "Level"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/practices": {
      "post": {
        "tags": [
          "Practice"
        ],
        "summary": "Create Practice",
        "operationId": "create_practice_api_v1_practices_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PracticeCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Practice"
        ],
        "summary": "List Practices",
        "operationId": "list_practices_api_v1_practices_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(active|completed|abandoned)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticesOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/practices/{practice_id}": {
      "get": {
        "tags": [
          "Practice"
        ],
        "summary": "Get Practice",
        "operationId": "get_practice_api_v1_practices__practice_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "practice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Practice Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/exam-practice/exams": {
      "get": {
        "tags": [
          "Exam practice"
        ],
        "summary": "List Exams",
        "operationId": "list_exams_api_v1_exam_practice_exams_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "level",
            "in": "query",
            "required": true,
            "schema": {
              "enum": [
                "N2",
                "N3"
              ],
              "type": "string",
              "title": "Level"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": true,
            "schema": {
              "enum": [
                "vocabulary",
                "grammar",
                "reading",
                "listening"
              ],
              "type": "string",
              "title": "Category"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamPracticeList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/exam-practice/exams/{exam_id}/types": {
      "get": {
        "tags": [
          "Exam practice"
        ],
        "summary": "List Types",
        "operationId": "list_types_api_v1_exam_practice_exams__exam_id__types_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "exam_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Exam Id"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": true,
            "schema": {
              "enum": [
                "vocabulary",
                "grammar",
                "reading",
                "listening"
              ],
              "type": "string",
              "title": "Category"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamPracticeTypes"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/exam-practice/exams/{exam_id}/types/{type_id}/practice": {
      "post": {
        "tags": [
          "Exam practice"
        ],
        "summary": "Start Exam Type",
        "operationId": "start_exam_type_api_v1_exam_practice_exams__exam_id__types__type_id__practice_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "exam_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Exam Id"
            }
          },
          {
            "name": "type_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Type Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ExamOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "level": {
            "type": "string",
            "title": "Level"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          },
          "month": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Month"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "level",
          "year",
          "month"
        ],
        "title": "ExamOut"
      },
      "ExamPracticeList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ExamProgressOut"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ExamPracticeList"
      },
      "ExamPracticeTypes": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ExamTypeProgressOut"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ExamPracticeTypes"
      },
      "ExamProgressOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "level": {
            "type": "string",
            "title": "Level"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          },
          "month": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Month"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "answered": {
            "type": "integer",
            "title": "Answered"
          },
          "correct": {
            "type": "integer",
            "title": "Correct"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "level",
          "year",
          "month",
          "total",
          "answered",
          "correct",
          "status"
        ],
        "title": "ExamProgressOut"
      },
      "ExamTypeProgressOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name_zh": {
            "type": "string",
            "title": "Name Zh"
          },
          "name_ja": {
            "type": "string",
            "title": "Name Ja"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "answered": {
            "type": "integer",
            "title": "Answered"
          },
          "correct": {
            "type": "integer",
            "title": "Correct"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "practice_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Practice Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name_zh",
          "name_ja",
          "category",
          "total",
          "answered",
          "correct",
          "status",
          "practice_id"
        ],
        "title": "ExamTypeProgressOut"
      },
      "ExamsOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ExamOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "ExamsOut"
      },
      "FeedbackOut": {
        "properties": {
          "chosen_option_id": {
            "type": "string",
            "title": "Chosen Option Id"
          },
          "correct_option_id": {
            "type": "string",
            "title": "Correct Option Id"
          },
          "is_correct": {
            "type": "boolean",
            "title": "Is Correct"
          },
          "explanation": {
            "$ref": "#/components/schemas/RichText"
          },
          "explanation_available": {
            "type": "boolean",
            "title": "Explanation Available"
          },
          "translation": {
            "$ref": "#/components/schemas/RichText"
          },
          "subtitles": {
            "items": {
              "$ref": "#/components/schemas/SubtitleOut"
            },
            "type": "array",
            "title": "Subtitles"
          },
          "answered_at": {
            "type": "string",
            "title": "Answered At"
          },
          "elapsed_ms": {
            "type": "integer",
            "title": "Elapsed Ms"
          }
        },
        "type": "object",
        "required": [
          "chosen_option_id",
          "correct_option_id",
          "is_correct",
          "explanation",
          "explanation_available",
          "translation",
          "subtitles",
          "answered_at",
          "elapsed_ms"
        ],
        "title": "FeedbackOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ItemOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "question": {
            "$ref": "#/components/schemas/QuestionOut"
          },
          "feedback": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FeedbackOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "position",
          "question",
          "feedback"
        ],
        "title": "ItemOut"
      },
...
      "MaterialOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "content": {
            "$ref": "#/components/schemas/RichText"
          },
          "audio_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Url"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "content",
          "audio_url",
          "image_url"
        ],
        "title": "MaterialOut"
      },
      "OptionOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "content": {
            "$ref": "#/components/schemas/RichText"
          }
        },
        "type": "object",
        "required": [
          "id",
          "position",
          "content"
        ],
        "title": "OptionOut"
      },
      "PracticeCreate": {
        "properties": {
          "level": {
            "type": "string",
            "enum": [
              "N2",
              "N3"
            ],
            "title": "Level"
          },
          "type_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Type Id"
          },
          "count": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Count",
            "default": 10
          },
          "mode": {
            "type": "string",
            "enum": [
              "normal",
              "wrong"
            ],
            "title": "Mode",
            "default": "normal"
          },
          "request_key": {
            "type": "string",
            "maxLength": 64,
            "minLength": 8,
            "pattern": "^[a-zA-Z0-9_-]+$",
            "title": "Request Key"
          }
        },
        "type": "object",
        "required": [
          "level",
          "type_id",
          "request_key"
        ],
        "title": "PracticeCreate"
      },
      "PracticeOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "level": {
            "type": "string",
            "title": "Level"
          },
          "type_id": {
            "type": "string",
            "title": "Type Id"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "requested_count": {
            "type": "integer",
            "title": "Requested Count"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "answered": {
            "type": "integer",
            "title": "Answered"
          },
          "correct": {
            "type": "integer",
            "title": "Correct"
          },
          "elapsed_ms": {
            "type": "integer",
            "title": "Elapsed Ms"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "next_item_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Item Id"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ItemOut"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "id",
          "level",
          "type_id",
          "mode",
          "status",
          "requested_count",
          "total",
          "answered",
          "correct",
          "elapsed_ms",
          "created_at",
          "completed_at",
          "next_item_id",
          "items"
        ],
        "title": "PracticeOut"
      },
      "PracticeSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "level": {
            "type": "string",
            "title": "Level"
          },
          "type_id": {
            "type": "string",
            "title": "Type Id"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "requested_count": {
            "type": "integer",
            "title": "Requested Count"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "answered": {
            "type": "integer",
            "title": "Answered"
          },
          "correct": {
            "type": "integer",
            "title": "Correct"
          },
          "elapsed_ms": {
            "type": "integer",
            "title": "Elapsed Ms"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "next_item_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Item Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "level",
          "type_id",
          "mode",
          "status",
          "requested_count",
          "total",
          "answered",
          "correct",
          "elapsed_ms",
          "created_at",
          "completed_at",
          "next_item_id"
        ],
        "title": "PracticeSummary"
      },
      "PracticesOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PracticeSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "PracticesOut"
      },
...
      "QuestionOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "occurrence_id": {
            "type": "string",
            "title": "Occurrence Id"
          },
          "type_id": {
            "type": "string",
            "title": "Type Id"
          },
          "group_id": {
            "type": "string",
            "title": "Group Id"
          },
          "prompt": {
            "$ref": "#/components/schemas/RichText"
          },
          "material": {
            "$ref": "#/components/schemas/MaterialOut"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/OptionOut"
            },
            "type": "array",
            "title": "Options"
          },
          "source": {
            "$ref": "#/components/schemas/SourceOut"
          }
        },
        "type": "object",
        "required": [
          "id",
          "occurrence_id",
          "type_id",
          "group_id",
          "prompt",
          "material",
          "options",
          "source"
        ],
        "title": "QuestionOut"
      },
      "RichText": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "html": {
            "type": "string",
            "title": "Html"
          }
        },
        "type": "object",
        "required": [
          "text",
          "html"
        ],
        "title": "RichText"
      },
      "SourceOut": {
        "properties": {
          "exam_id": {
            "type": "string",
            "title": "Exam Id"
          },
          "exam_title": {
            "type": "string",
            "title": "Exam Title"
          },
          "level": {
            "type": "string",
            "title": "Level"
          },
          "section": {
            "$ref": "#/components/schemas/RichText"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          }
        },
        "type": "object",
        "required": [
          "exam_id",
          "exam_title",
          "level",
          "section",
          "position"
        ],
        "title": "SourceOut"
      },
...
      "SubtitleOut": {
        "properties": {
          "start_ms": {
            "type": "integer",
            "title": "Start Ms"
          },
          "end_ms": {
            "type": "integer",
            "title": "End Ms"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "start_ms",
          "end_ms",
          "text"
        ],
        "title": "SubtitleOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}